Preserve the battle setup slot index width
Model the inner battle roster slot as the retail u16 induction variable so its truncation, unsigned bounds, and wrap behavior compile into the target control shape.

Constraint: Keep the outer player index signed and unchanged.

Confidence: high

Scope-risk: narrow

Tested: ninja -j4 build/GC6E01/src/game/menu/menuCB_BattleStart_candidate_8005FFE4.o build/GC6E01/report.json; function fuzzy 39.97826% -> 43.072464%, unit 57.040455% -> 57.92635%.

+++ src/game/menu/menuCB_BattleStart.c
@@ -635,7 +635,7 @@ void _menuCBBattleStartSetIndex__Fv(void)
     u16 count[4];
     s32 battleType;
     s32 player;
-    s32 slot;
+    u16 slot;
 
     order = (u32 (*)[6])lbl_803A9E40;
     battleType = toolentryTaisenGetBattleType();
